Moved the core/testing file into the tests/core module
