+++ tests/api/admin/test_geographic_validator.py
@@ -10,7 +10,7 @@ from api.admin.geographic_validator import GeographicValidator
 from api.admin.problem_details import *
 from api.registration.registry import RemoteRegistry
 from core.model import ExternalIntegration, create
-from core.testing import MockRequestsResponse
+from tests.core.mock import MockRequestsResponse
 
 
 class TestGeographicValidator:
