+++ tests/api/conftest.py
@@ -5,7 +5,6 @@ register_assert_rewrite("tests.fixtures.files")
 register_assert_rewrite("tests.fixtures.vendor_id")
 
 pytest_plugins = [
-    "core.testing",
     "tests.fixtures.announcements",
     "tests.fixtures.api_admin",
     "tests.fixtures.api_axis_files",
