+++ tests/api/mockapi/axis.py
@@ -45,7 +45,7 @@ class MockAxis360API(Axis360API):
         self.requests = []
 
     def queue_response(self, status_code, headers={}, content=None):
-        from core.testing import MockRequestsResponse
+        from tests.core.mock import MockRequestsResponse
 
         self.responses.insert(0, MockRequestsResponse(status_code, headers, content))
