+++ tests/api/mockapi/bibliotheca.py
@@ -4,8 +4,8 @@ from api.bibliotheca import BibliothecaAPI
 from core.model import get_one_or_create
 from core.model.collection import Collection
 from core.model.configuration import ExternalIntegration
-from core.testing import MockRequestsResponse
 from core.util.http import HTTP
+from tests.core.mock import MockRequestsResponse
 from tests.fixtures.db import DatabaseTest
