+++ tests/fixtures/odl.py
@@ -20,8 +20,8 @@ from core.model import (
     get_one_or_create,
 )
 from core.model.configuration import ExternalIntegration
-from core.testing import MockRequestsResponse
 from core.util.http import HTTP
+from tests.core.mock import MockRequestsResponse
 from tests.fixtures.api_odl2_files import ODL2APIFilesFixture
 from tests.fixtures.api_odl_files import ODLAPIFilesFixture
 from tests.fixtures.database import DatabaseTransactionFixture
